Added RecAny to allow recursive concept types.
Alright, so this is what's going on:

I renamed `Any` to `RecAny` and removed its `final` specifier.
Then I created a new `Any` final class that inherits from `RecAny`.

The new `Any` class should be 100% compatible with old code.

The new `RecAny` class is also the same, but since it's not final,
    it can be used in CRTP patterns.